Practical 8 Part 2
Reads from a file. Makes things slightly easier.
Unfortunately everything needs to be in a line. A loop might be needed to read multiple lines.

There are also minor problems. All punctuation are ignored, including ' (which creates s key with value of 4) and . in e.g. (which creates a weird e and g key)
Excluding ' might work, but it will create other weird situations where quotes in ' would have ' in front and behind specific words. 
Excluding . will not work, as they are far too common. It can be possible to exclude the specific word e.g., but it only works in this specific case. In a realistic case, it will fail because some other combination might exist, and things like this should not be done in a case-by-case basis.