Practical 8 Part 3
Is the count not needed? If it isnt, the code would definitely be far more cleaner than it is. 

Using an ArrayList, 2 of them, actually, i stored a list of files and a list of hashmaps. Each file has an hashmap associated to it, albeit a little rough. It might be better if i created an object for each of them, which contains a file and a hashmap for a cleaner result. Regardless, it works for now. 

I've added a hmTotal but it seems like it isnt needed. Still, it can be used as a contingency?

The printing part in the end is definitely not done, however, and would need some work.